Refactor test files and update dependencies
