--- tests/api.test.ts
@@ -1,65 +1,0 @@
-import { assert, assertThrows } from 'https://deno.land/std@0.219.0/assert/mod.ts';
-
-import { RULE } from '../api/api.ts';
-
-Deno.test('validates a check rule', () => {
-  const dummyRule = RULE.parse({
-    check: async () => {},
-  });
-  assert(dummyRule);
-});
-
-Deno.test('validates a progress rule', () => {
-  const dummyRule = RULE.parse({
-    progress: async () => {},
-  });
-  assert(dummyRule);
-});
-
-Deno.test('validates an action rule', () => {
-  const dummyRule = RULE.parse({
-    action: async () => {},
-  });
-  assert(dummyRule);
-});
-
-Deno.test('validates a combined rule', async (t) => {
-  await t.step('A & B', () => {
-    const dummyRule = RULE.parse({
-      action: async () => {},
-      check: async () => {},
-    });
-    assert(dummyRule);
-  });
-
-  await t.step('A & C', () => {
-    const dummyRule = RULE.parse({
-      action: async () => {},
-      check: async () => {},
-    });
-    assert(dummyRule);
-  });
-
-  await t.step('B & C', () => {
-    const dummyRule = RULE.parse({
-      action: async () => {},
-      check: async () => {},
-    });
-    assert(dummyRule);
-  });
-
-  await t.step('A & B & C', () => {
-    const dummyRule = RULE.parse({
-      action: async () => {},
-      check: async () => {},
-      progress: async () => {},
-    });
-    assert(dummyRule);
-  });
-});
-
-Deno.test('rejects invalid rule', () => {
-  assertThrows(() => {
-    RULE.parse({});
-  });
-});
